Complete search logic extraction to useSearch hook
