+++ src/components/HomePage.jsx
@@ -10,14 +10,50 @@ import useSearch from '../hooks/useSearch.js';
  * @returns {JSX.Element} The HomePage component
  */
 function HomePage() {
-  // Use search hook for all search functionality
-  const { searchResults, isSearching, searchError, currentQuery, showResults, search, lookAhead } = useSearch();
+  // Use search hook for all search functionality - single source of truth
+  const {
+    // Main search state
+    searchResults,
+    isSearching,
+    searchError,
+    currentQuery,
+    showResults,
+    
+    // Input and suggestions state
+    inputValue,
+    suggestions,
+    showDropdown,
+    isLoadingSuggestions,
+    
+    // Refs
+    dropdownRef,
+    
+    // Action functions
+    handleInputChange,
+    handleSubmit,
+    handleSuggestionClick,
+    handleResultClick,
+    handleClear
+  } = useSearch();
+
+  // Props to pass to SearchBar via SearchHero
+  const searchProps = {
+    inputValue,
+    suggestions,
+    showDropdown,
+    isLoadingSuggestions,
+    dropdownRef,
+    handleInputChange,
+    handleSubmit,
+    handleSuggestionClick,
+    handleClear
+  };
 
   return (
     <div className="min-h-screen hero-gradient">
       <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
         {/* Hero Section with Search */}
-        <SearchHero onSearch={search} onLookAhead={lookAhead} />
+        <SearchHero searchProps={searchProps} />
 
         {/* Search Results */}
         {showResults && (
@@ -27,6 +63,7 @@ function HomePage() {
               isLoading={isSearching}
               error={searchError}
               query={currentQuery}
+              onResultClick={handleResultClick}
             />
           </section>
         )}
